refactor: simplify trust policy plumbing
