+++ packages/provision-kit/src/install-source-download.ts
@@ -64,6 +64,7 @@ async function requestHop(
       family: approved.family,
       headers: { ...headers, 'accept-encoding': 'identity' },
       signal,
+      method: 'GET',
     });
   } catch (error) {
     if (error instanceof AppError) throw error;
